requirements.txt
rich
requests
google-generativeai
openai
anthropic
mistralai
pyyaml
fake-useragent
beautifulsoup4